% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dbar.R
\name{convert_dbar}
\alias{convert_dbar}
\title{convert_dbar}
\usage{
convert_dbar(val, to_unit = "m")
}
\arguments{
\item{val}{value in dbar to convert}

\item{to_unit}{abbreviation for unit ('m', 'ft')}
}
\value{
data.table of results
}
\description{
convert values from dbar
}
\author{
Jonathan Kennel \email{jkennel@uoguelph.ca}
}
