% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fst.R
\name{filter_fst}
\alias{filter_fst}
\title{Obtain data from fst file subset by dates and well name}
\usage{
filter_fst(db_name, start_date, end_date, well_names = NA)
}
\arguments{
\item{db_name}{the path to the rbr database ( rsk )}

\item{start_date}{character date}

\item{end_date}{character date}

\item{well_names}{character specific wells to subset}
}
\value{
data.table of results
}
\description{
Import fst data to R
}
\author{
Jonathan Kennel \email{jkennel@uoguelph.ca}
}
