% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rbr.R
\name{filter_rbr}
\alias{filter_rbr}
\title{obtain data from rbr sqlite3 database}
\usage{
filter_rbr(db_name, start_date, end_date, tz = "UTC")
}
\arguments{
\item{db_name}{the path to the rbr database ( rsk )}

\item{start_date}{character date}

\item{end_date}{character date}

\item{tz}{the timezone of the input data file}
}
\value{
data.table of results
}
\description{
import sqlite data to R
}
\author{
Jonathan Kennel \email{jkennel@uoguelph.ca}
}

