% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rbr_files.R
\name{filter_rbr_files}
\alias{filter_rbr_files}
\title{obtain data from folder containing rbr sqlite3 databases}
\usage{
filter_rbr_files(files, start_date, end_date, tz = "UTC")
}
\arguments{
\item{files}{the path to the rbr database ( rsk )}

\item{start_date}{character date}

\item{end_date}{character date}

\item{use_rbr_tz}{boolean use time zone information from the rbr file?}
}
\value{
data.table of results
}
\description{
import sqlite data to R
}
\author{
Jonathan Kennel \email{jkennel@uoguelph.ca}
}
