% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rbr_folder.R
\name{filter_rbr_folder}
\alias{filter_rbr_folder}
\title{obtain data from folder containing rbr sqlite3 databases}
\usage{
filter_rbr_folder(folder_path, start_date, end_date, use_rbr_tz = TRUE, ...)
}
\arguments{
\item{folder_path}{the path to the rbr database ( rsk )}

\item{start_date}{character date}

\item{end_date}{character date}

\item{use_rbr_tz}{boolean use time zone information from the rbr file?}

\item{...}{arguments to pass to list.files}
}
\value{
data.table of results
}
\description{
import sqlite data to R
}
\author{
Jonathan Kennel \email{jkennel@uoguelph.ca}
}
