% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dates.R
\name{find_nearest}
\alias{find_nearest}
\title{find_nearest}
\usage{
find_nearest(transducer, manual, roll_size = 86400 * 7)
}
\arguments{
\item{transducer}{data.table of transducer measurements (name, datetime)}

\item{manual}{data.table manual measurements (name, datetime)}

\item{roll_size}{numeric maximum time difference (sec) to allow}
}
\value{
transducer measurements within roll_size time
}
\description{
find the closest value in time
}
\author{
Jonathan Kennel \email{jkennel@uoguelph.ca}
}
