% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbr_tables.R
\name{int64_to_posix}
\alias{int64_to_posix}
\title{int64_to_posix}
\usage{
int64_to_posix(dt)
}
\arguments{
\item{dt}{the table of data}
}
\value{
data.table with integer64 converted to POSIXct
}
\description{
convert integer64 to posix
}
\author{
Jonathan Kennel \email{jkennel@uoguelph.ca}
}
