% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_log_start.R
\name{parse_log}
\alias{parse_log}
\title{parse_log_start}
\usage{
parse_log(file)
}
\arguments{
\item{file}{input file}
}
\value{
starting info
}
\description{
Get transducer info from rbr log file
}
\author{
Jonathan Kennel \email{jkennel@uoguelph.ca}
}
