% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rbr_download_date.R
\name{parse_rbr_download_date}
\alias{parse_rbr_download_date}
\title{parse_rbr_download_date}
\usage{
parse_rbr_download_date(fn, tz = "UTC")
}
\arguments{
\item{fn}{{character} scalar or vector of file name(s)}

\item{tz}{{character} timezone}
}
\value{
POSIXct datetime
}
\description{
RBR files typically save the date in the downloaded file name.  This function
parses that date to a POSIXct
}
\examples{
dt <- parse_rbr_download_date('20141117_1240')
dt <- parse_rbr_download_date('201adf41a11a7_1240')
}
