% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_log_start.R
\name{parse_start}
\alias{parse_start}
\title{parse_start}
\usage{
parse_start(search_string, x)
}
\arguments{
\item{search_string}{text identifier}

\item{x}{text}
}
\value{
starting info
}
\description{
parse the starting info from an rbr log file
}
\author{
Jonathan Kennel \email{jkennel@uoguelph.ca}
}
