% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbr_info.R
\name{rbr_info}
\alias{rbr_info}
\title{obtain meta data from rbr sqlite3 database}
\usage{
rbr_info(db_name, sql_text, tz = "UTC")
}
\arguments{
\item{db_name}{character the path to the rbr database ( rsk )}
}
\value{
data.table of results
}
\description{
get supplementary info from .rsk file
}
\author{
Jonathan Kennel \email{jkennel@uoguelph.ca}
}
