% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbr_tables.R
\name{rbr_table_names}
\alias{rbr_table_names}
\title{rbr_table_names}
\usage{
rbr_table_names(db_name)
}
\arguments{
\item{db_name}{the path to the rbr database ( rsk )}
}
\value{
vector of table names
}
\description{
get the names of the tables in the .rsk file
}
\author{
Jonathan Kennel \email{jkennel@uoguelph.ca}
}
