% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbr_tables.R
\name{rbr_tables}
\alias{rbr_tables}
\title{rbr_tables}
\usage{
rbr_tables(db_name)
}
\arguments{
\item{db_name}{the path to the rbr database ( rsk )}
}
\value{
the .rsk file as a list of tables
}
\description{
get the raw tables from the .rsk file
}
\author{
Jonathan Kennel \email{jkennel@uoguelph.ca}
}
