% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fst.R
\name{read_fst}
\alias{read_fst}
\title{Obtain data from fst file}
\usage{
read_fst(db_name)
}
\arguments{
\item{db_name}{the path to the rbr database ( rsk )}
}
\value{
data.table of results
}
\description{
Import fst data to R
}
\author{
Jonathan Kennel \email{jkennel@uoguelph.ca}
}
