% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rbr.R
\name{read_rbr}
\alias{read_rbr}
\title{obtain data from rbr sqlite3 database}
\usage{
read_rbr(db_name, tz = "UTC")
}
\arguments{
\item{db_name}{the path to the rbr database ( rsk )}

\item{tz}{the timezone of the input data file}
}
\value{
data.table of results
}
\description{
import sqlite data to R
}
\author{
Jonathan Kennel \email{jkennel@uoguelph.ca}
}

