% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rbr_db.R
\name{read_rbr_db}
\alias{read_rbr_db}
\title{obtain data from rbr sqlite3 database}
\usage{
read_rbr_db(db_name, sql_text, tz = "UTC")
}
\arguments{
\item{db_name}{character the path to the rbr database ( rsk )}

\item{sql_text}{character sql string to execute on connection for filtering}

\item{tz}{character the timezone of the input data file}
}
\value{
data.table of results
}
\description{
import sqlite data to R
}
\author{
Jonathan Kennel \email{jkennel@uoguelph.ca}
}
