% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rbr_folder.R
\name{read_rbr_folder}
\alias{read_rbr_folder}
\title{obtain data from folder containing rbr sqlite3 databases}
\usage{
read_rbr_folder(folder_path, tz = "UTC", ...)
}
\arguments{
\item{folder_path}{the path to the rbr database ( rsk )}

\item{tz}{the timezone of the input data file}

\item{...}{arguments to pass to list.files}
}
\value{
data.table of results
}
\description{
import sqlite data to R
}
\author{
Jonathan Kennel \email{jkennel@uoguelph.ca}
}
