% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_data.R
\name{reference_data}
\alias{reference_data}
\title{reference_data}
\usage{
reference_data(data_depth, id_col = "id", path_col = "path",
  depth_col = "depth", type_col = "baro", elev_unit = "m",
  ref_elev = 0, start_date, end_date)
}
\arguments{
\item{data_depth}{table with id, path, and depth of the transducers}

\item{id_col}{column with the transducer id}

\item{path_col}{column with the path to the .rsk file}

\item{depth_col}{depth of the transducer from reference}

\item{type_col}{type of data being collected}

\item{elev_unit}{reference elevation units}

\item{ref_elev}{elevation of the reference}

\item{start_date}{starting date for the reference}

\item{end_date}{end date for the reference}
}
\description{
Convert pressure measurements to elevation, depth to water and ensure
consistent units
}
