% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dates.R
\name{shift_values_range}
\alias{shift_values_range}
\title{shift_values_range}
\usage{
shift_values_range(dat, shift)
}
\arguments{
\item{dat}{data.table to filter (name, datetime)}

\item{shift}{data.table of filter start and end times (start, end)}
}
\value{
data.table with shifted values
}
\description{
shift values for specific regions
}
\author{
Jonathan Kennel \email{jkennel@uoguelph.ca}
}
