% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dates.R
\name{filter_dates}
\alias{filter_dates}
\title{filter_dates}
\usage{
filter_dates(dat, filt, keep = FALSE, which = FALSE,
  include_filt_cols = FALSE)
}
\arguments{
\item{dat}{data.table to filter (name, datetime)}

\item{filt}{data.table of filter start and end times (start, end)}

\item{keep}{include or exclude the subsets}

\item{which}{return the indices that match}

\item{include_filt_cols}{include the columns in the filter table}
}
\value{
filtered data.table
}
\description{
remove values from a data.table (non-equi join filter)
}
\author{
Jonathan Kennel \email{jkennel@uoguelph.ca}
}
