% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_diver.R
\name{read_diver}
\alias{read_diver}
\title{obtain data from diver .mon file}
\usage{
read_diver(db_name, tz = "UTC")
}
\arguments{
\item{db_name}{character the path to the rbr database ( rsk )}

\item{use_rbr_tz}{boolean use time zone information from the rbr file?}
}
\value{
data.table of results
}
\description{
import diver data to R
}
\author{
Jonathan Kennel \email{jkennel@uoguelph.ca}
}
