% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dates.R
\name{shift_values_range}
\alias{shift_values_range}
\title{shift_values_range}
\usage{
shift_values_range(dat, shift)
}
\arguments{
\item{dat}{data.table to filter (name, datetime)}

\item{shift}{data.table of filter start and end times (start, end)}
}
\value{
data.table with shifted values
}
\description{
shift values for specific regions
}
\details{
#===============================================================================
#' @title filter_dates
#'
#' @description remove values from a data.table (non-equi join filter)
#'
#' @author Jonathan Kennel \email{jkennel@uoguelph.ca}
#'
#' @param dat data.table to filter (name, datetime)
#' @param filt data.table of filter start and end times (start, end)
#'
#' @return filtered data.table
#'
#' @export
#===============================================================================
filter_dates <- function(dat, filt){

dat[!J(filt), on = .(name, datetime >= start, datetime <= end)]

}
}
\author{
Jonathan Kennel \email{jkennel@uoguelph.ca}
}
